chore: move template filter to a dedicated place
